+++ CHANGES.rst
@@ -17,6 +17,7 @@ develop
 - hiveeyes: receive messages via MQTT and store data points into InfluxDB
 - app: use configuration file instead of hardcoded configuration values
 - storage: add support for InfluxDB 0.9
+- ui: add pages with page transitions, about content, etc.
 
 
 2015-05-21 0.2.2